redid routes structure
