+++ server/routes/middleware/auth.ts
